+++ cli/src/account.rs
@@ -107,3 +107,23 @@ pub fn operations(client: &Client, matches: &ArgMatches) -> Result<()> {
     });
     res
 }
+
+pub fn payments(client: &Client, matches: &ArgMatches) -> Result<()> {
+    let pager = Pager::from_arg(&matches);
+
+    let id = matches.value_of("ID").expect("ID is required");
+    let endpoint = account::Payments::new(id);
+    let endpoint = pager.assign(endpoint);
+    let endpoint = cursor::assign_from_arg(matches, endpoint);
+    let endpoint = ordering::assign_from_arg(matches, endpoint);
+
+    let iter = sync::Iter::new(&client, endpoint);
+
+    let mut res = Ok(());
+    let mut fmt = Formatter::start_stdout(Simple::new());
+    pager.paginate(iter, |result| match result {
+        Ok(op) => fmt.render(&op),
+        Err(err) => res = Err(err.into()),
+    });
+    res
+}