+++ cli/src/main.rs
@@ -131,6 +131,18 @@ fn build_app<'a, 'b>() -> App<'a, 'b> {
                             )
                     ),
 
+                )
+                .subcommand(
+                    listable!(
+                        SubCommand::with_name("payments")
+                            .about("Fetch all payments associated with an account")
+                            .arg(
+                                Arg::with_name("ID")
+                                    .required(true)
+                                    .help("The identifier of the account to look up"),
+                            )
+                    ),
+
                 ),
         )
         .subcommand(
@@ -440,6 +452,7 @@ fn main() {
             ("effects", Some(sub_m)) => account::effects(&client, sub_m),
             ("offers", Some(sub_m)) => account::offers(&client, sub_m),
             ("operations", Some(sub_m)) => account::operations(&client, sub_m),
+            ("payments", Some(sub_m)) => account::payments(&client, sub_m),
             _ => return print_help_and_exit(),
         },
         ("assets", Some(sub_m)) => match sub_m.subcommand() {